+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## [Unreleased]
 ### Added
-- User settings by group
+- Option to allow users to choose a SecSign ID during onboarding
+- Filter users in User Management by group or a search term
+- Sort users table
+- Slight UI and wording improvements
+
 
 ## 0.2.2 - 2019-05-22
 ### Added